pyproject.toml
[project]
name = "smart-captureplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
    "azure-storage-blob>=12.26.0",
    "openai>=1.99.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
